rename readOnly to constget
